requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-dotenv==1.0.1
feedparser==6.0.11
httpx==0.27.2
sqlalchemy==2.0.36
apscheduler==3.10.4
pytz>=2024.1
aiohttp==3.11.10
beautifulsoup4==4.12.3
lxml==5.3.0
